Changed City.id and City.parent_id type to long long

+++ db/citiesrepository.cpp
@@ -108,7 +108,7 @@ QList<Type> CitiesRepository::getTypes() {
   return result;
 }
 
-void CitiesRepository::deleteCity(qlonglong cityId) {
+void CitiesRepository::deleteCity(long long cityId) {
   QSqlQuery query;
   query.prepare("DELETE FROM cities WHERE id = :id");
   query.bindValue(":id", cityId);
@@ -137,11 +137,11 @@ void CitiesRepository::updateCity(City city) {
   query.bindValue(":map_point", city.m_map_point);
   query.bindValue(":lat", city.m_lat);
   query.bindValue(":lon", city.m_lon);
-  query.bindValue(":alt", city.m_alt); // TODO: long problem
+  query.bindValue(":alt", city.m_alt);
   query.bindValue(":population", city.m_population);
   query.bindValue(":description", city.m_description);
   query.bindValue(":country", city.m_country);
-  query.bindValue(":id", static_cast<qlonglong>(city.m_id));
+  query.bindValue(":id", city.m_id);
 
   if (!query.exec()) {
     qWarning() << "Error while exec query:" << query.lastError().text();
